refactor(back-end): split endpoints in different files
